+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * feat: Add `air_public_input` flag to `cairo1-run` [#1539] (https://github.com/lambdaclass/cairo-vm/pull/1539)
 
+* feat: Implement air_private_input [#1552](https://github.com/lambdaclass/cairo-vm/pull/1552)
+
 * feat: Add `proof_mode` flag to `cairo1-run` [#1537] (https://github.com/lambdaclass/cairo-vm/pull/1537)
   * The cairo1-run crate no longer compiles and executes in proof_mode by default
   * Add flag `proof_mode` to cairo1-run crate. Activating this flag will enable proof_mode compilation and execution